Rename folder and add VPC id export
