+++ lib/components/vpc-details.ts
